Make Redis tests abstract
